Group privileges definitions in a single file
