--- specification/security/_types/ApplicationPrivileges.ts
@@ -1,33 +1,0 @@
-/*
- * Licensed to Elasticsearch B.V. under one or more contributor
- * license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright
- * ownership. Elasticsearch B.V. licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-export class ApplicationPrivileges {
-  /**
-   * The name of the application to which this entry applies.
-   */
-  application: string
-  /**
-   * A list of strings, where each element is the name of an application privilege or action.
-   */
-  privileges: string[]
-  /**
-   * A list resources to which the privileges are applied.
-   */
-  resources: string[]
-}
