+++ specification/security/_types/Privileges.ts
@@ -17,9 +17,63 @@
  * under the License.
  */
 
-import { FieldSecurity } from '@security/_types/FieldSecurity'
-import { Indices } from '@_types/common'
-import { QueryContainer } from '@_types/query_dsl/abstractions'
+import { Indices } from "@_types/common"
+import { QueryContainer } from "@_types/query_dsl/abstractions"
+import { FieldSecurity } from "./FieldSecurity"
+
+export class ApplicationPrivileges {
+  /**
+   * The name of the application to which this entry applies.
+   */
+  application: string
+  /**
+   * A list of strings, where each element is the name of an application privilege or action.
+   */
+  privileges: string[]
+  /**
+   * A list resources to which the privileges are applied.
+   */
+  resources: string[]
+}
+
+export enum ClusterPrivilege {
+  all,
+  cancel_task,
+  create_snapshot,
+  grant_api_key,
+  manage,
+  manage_api_key,
+  manage_ccr,
+  manage_ilm,
+  manage_index_templates,
+  manage_ingest_pipelines,
+  manage_logstash_pipelines,
+  manage_ml,
+  manage_oidc,
+  manage_own_api_key,
+  manage_pipeline,
+  manage_rollup,
+  manage_saml,
+  manage_security,
+  manage_service_account,
+  manage_slm,
+  manage_token,
+  manage_transform,
+  manage_watcher,
+  monitor,
+  monitor_ml,
+  monitor_rollup,
+  monitor_snapshot,
+  monitor_text_structure,
+  monitor_transform,
+  monitor_watcher,
+  read_ccr,
+  read_ilm,
+  read_pipeline,
+  read_slm,
+  transport_client
+}
+
 
 export class IndicesPrivileges {
   /**
@@ -65,4 +119,16 @@ export enum IndexPrivilege {
   read_cross_cluster,
   view_index_metadata,
   write
+}
+
+export class GlobalPrivilege {
+  application: ApplicationGlobalUserPrivileges
+}
+
+export class ApplicationGlobalUserPrivileges {
+  manage: ManageUserPrivileges
+}
+
+export class ManageUserPrivileges {
+  applications: string[]
 }
